merda

+++ src/main/java/com/mycompany/projetopoo/CISUC.java
@@ -222,7 +222,8 @@ public int leFicheiroProjetos(){
                     case "DOCU":{
                         DataI = (line.split("/")[2]).split("-");
                         DataF = (line.split("/")[4]).split("-");
-                        int taxa = Integer.parseInt(line.split("/")[6]);
+                        
+                        Double taxa = Double.parseDouble(line.split("/")[6]);
                         GregorianCalendar datai = new GregorianCalendar(Integer.parseInt(DataI[2]),Integer.parseInt(DataI[1]),Integer.parseInt(DataI[0]));
                         if(CheckdataGregorian(datai)==1){
                             System.out.println("DATA INICIAL INVALIDA NA TAREFA DE DOCUMENTACAO "+line.split("/")[1]+": LEITURA DO FICHEIRO");
@@ -251,7 +252,7 @@ public int leFicheiroProjetos(){
                     case "DESI":{
                         DataI = (line.split("/")[2]).split("-");
                         DataF = (line.split("/")[4]).split("-");
-                        int taxa = Integer.parseInt(line.split("/")[6]);
+                        double taxa = Double.parseDouble(line.split("/")[6]);
                         GregorianCalendar datai = new GregorianCalendar(Integer.parseInt(DataI[2]),Integer.parseInt(DataI[1]),Integer.parseInt(DataI[0]));
                         if(CheckdataGregorian(datai)==1){
                             System.out.println("DATA INICIAL INVALIDA NA TAREFA DE Design "+line.split("/")[1]+": LEITURA DO FICHEIRO");
@@ -280,7 +281,7 @@ public int leFicheiroProjetos(){
                     case "DESE":{
                         DataI = (line.split("/")[2]).split("-");
                         DataF = (line.split("/")[4]).split("-");
-                        int taxa = Integer.parseInt(line.split("/")[6]);
+                        double taxa = Double.parseDouble(line.split("/")[6]);
                         GregorianCalendar datai = new GregorianCalendar(Integer.parseInt(DataI[2]),Integer.parseInt(DataI[1]),Integer.parseInt(DataI[0]));
                         if(CheckdataGregorian(datai)==1){
                             System.out.println("DATA INICIAL INVALIDA NA TAREFA DE DESENVOLVIMENTO "+line.split("/")[1]+": LEITURA DO FICHEIRO");
